+++ core/src/labels/labelMesh.cpp
@@ -104,7 +104,7 @@ void LabelMesh::draw(ShaderProgram& _shader) {
 
     size_t vertexOffset = 0;
 
-    for (int i = 0; i < m_vertexOffsets.size(); ++i) {
+    for (size_t i = 0; i < m_vertexOffsets.size(); ++i) {
         auto& o = m_vertexOffsets[i];
         uint32_t nIndices = o.first;
         uint32_t nVertices = o.second;
